+++ branch/src/main/scala/dev/wishingtree/branch/mustachio/MustachioDelimeteresSpec.test.scala
@@ -1,15 +1,8 @@
 package dev.wishingtree.branch.mustachio
 
-import scala.io.Source
-import scala.util.Using
+class MustachioDelimeteresSpec extends MustacheSpecSuite {
 
-class MustachioDelimeteresSpec extends MustachePartialsSpecSuite {
-
-  val specSuite: SpecSuite =
-    Using(Source.fromResource("mustache/delimeters.json")) { source =>
-      SpecSuite.decoder.decode(source.mkString)
-    }.flatten.getOrElse(throw new Exception("Failed to parse json"))
-
-  specSuite.tests.foreach(runSpec)
+  specSuite("mustache/delimiters.json").tests
+    .foreach(runSpec)
 
 }
