+++ branch/src/main/scala/dev/wishingtree/branch/mustachio/MustachioInterpolationSpec.test.scala
@@ -1,15 +1,8 @@
 package dev.wishingtree.branch.mustachio
 
-import scala.io.Source
-import scala.util.Using
-
 class MustachioInterpolationSpec extends MustacheSpecSuite {
 
-  val specSuite: SpecSuite =
-    Using(Source.fromResource("mustache/interpolation.json")) { source =>
-      SpecSuite.decoder.decode(source.mkString)
-    }.flatten.getOrElse(throw new Exception("Failed to parse json"))
-
-  specSuite.tests.foreach(runSpec)
+  specSuite("mustache/interpolation.json").tests
+    .foreach(runSpec)
 
 }
