+++ branch/src/main/scala/dev/wishingtree/branch/mustachio/MustachioPartialsSpec.test.scala
@@ -1,15 +1,8 @@
 package dev.wishingtree.branch.mustachio
 
-import scala.io.Source
-import scala.util.Using
+class MustachioPartialsSpec extends MustacheSpecSuite {
 
-class MustachioPartialsSpec extends MustachePartialsSpecSuite {
-
-  val specSuite: SpecSuite =
-    Using(Source.fromResource("mustache/partials.json")) { source =>
-      SpecSuite.decoder.decode(source.mkString)
-    }.flatten.getOrElse(throw new Exception("Failed to parse json"))
-
-  specSuite.tests.foreach(runSpec)
+  specSuite("mustache/partials.json").tests
+    .foreach(runSpec)
 
 }
